simplify show wdpa component
