--- app/layout/scenarios/show/wdpa/content/index.ts
@@ -1,1 +1,0 @@
-export { default } from './component';
